updated type to brand in routes and controllers
