--- routes/typeRoute.js
@@ -1,8 +1,0 @@
-const express = require('express')
-const router = express.Router()
-const controllers = require('../controllers/typeController')
-
-router.get('/', controllers.getAllTypes)
-router.get('/:id', controllers.getType)
-
-module.exports = router
